+++ CHANGELOG.md
@@ -39,7 +39,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - sails.helpers.break.addUnderwriting helper, with fastForwardOnly and quantity parameters. This helper determines which underwritings should be queued to air, and queues them.
  - Underwritings status system for alerting of underwritings that are behind schedule.
  - queueUnderwritings break task for specifying in breaks config to queue underwritings. The quantity parameter determines the maximum underwritings to queue (rule may be ignored if more underwritings are way behind schedule).
+ - Added Darksky model; fetches Darksky.net weather information for use on display signs and other applications via darksky/get and the darksky websocket event.
  - Added parameter "category" to songs/get; allows to filter songs by configured Node music category (not the same as RadioDJ main category).
+ - Added parameter "ignoreNonMusic" to songs/get. If true, non-music tracks will not be returned (eg. non-requestable tracks). Ignored if ID is specified.
+ - Added parameter "ignoreDisabled" to songs/get. If true, disabled tracks will not be returned. Ignored if ID is specified.
+ - Added parameter "ignoreSpins" to songs/get. If true, spin counts will not be returned, speeding up response time. Ignored if ID is not specified.
 
 ### Changed
  - Many of the methods used in api controllers and in models have been migrated to sails helpers.